Create 100 Doors
In this exercise I learned some valuable first steps to programming in Java. I reintroduced myself to nest for loops which bothered me the first time around but now fully grasp, the use of modulo as more than just a binary operator, and the use of boolean arrays and how they can easily be manipulated to switch with a ! (not) symbol.

I ran into trouble trying to figure out how I was going to get the program to realize that it needs to switch only certain doors after the first level of the for loop. Lucky for me, google exists and was able to explain the benefits of the modulo operator.