Create Creating Phone Numbers
There is no doubt that this could have been done better. But everything is a process and this was definitely a learning experience. On top of learning more about arrays, strings, and characters, I was able to effectively find information by navigating multiple stack overflow posts. This will prove to be a valuable skill to improve on due to the abundance of code I have yet to learn or understand.
There is a great feeling that comes with finding a piece of information you know in the back of your head is out there and you were looking for; that "Ah ha" moment.